feat(TDQ-17432): combined all masking components to use one migration
