+++ main/plugins/org.talend.repository/plugin.xml
@@ -3227,9 +3227,9 @@
         <projecttask
               beforeLogon="false"
               breaks="7.3.0"
-              class="org.talend.repository.model.migration.AddRAWFlowTPatternMasking"
+              class="org.talend.repository.model.migration.AddRAWFlowMaskingComponents"
               description="Add the check box to seperate the output"
-              id="org.talend.repository.model.migration.AddRAWFlowTPatternMasking"
+              id="org.talend.repository.model.migration.AddRAWFlowMaskingComponents"
               name="AddRAWFlowTPatternMasking"
               version="7.3.1">
         </projecttask>
@@ -3241,15 +3241,6 @@
               id="org.talend.repository.model.migration.changeDBVersionForMapROJAIOutputMigrationTask"
               name="changeDBVersionForMapROJAIOutputMigrationTask"
               version="7.0.1">
-        </projecttask>
-        <projecttask
-              beforeLogon="false"
-              breaks="7.3.0"
-              class="org.talend.repository.model.migration.AddRAWFlowTDataMasking"
-              description="Add the check box to seperate the output"
-              id="org.talend.repository.model.migration.AddRAWFlowTDataMasking"
-              name="AddRAWFlowTDataMasking"
-              version="7.3.1">
         </projecttask>
         <projecttask
               beforeLogon="false"
