+++ main/plugins/org.talend.repository/src/main/java/org/talend/repository/model/migration/AddRAWFlowMaskingComponents.java
@@ -35,8 +35,13 @@ import org.talend.designer.core.model.utils.emf.talendfile.MetadataType;
 import org.talend.designer.core.model.utils.emf.talendfile.NodeType;
 import org.talend.designer.core.model.utils.emf.talendfile.ProcessType;
 import org.talend.designer.core.model.utils.emf.talendfile.TalendFileFactory;
+import org.talend.designer.core.model.utils.emf.talendfile.impl.ConnectionTypeImpl;
+import org.talend.designer.core.model.utils.emf.talendfile.impl.NodeTypeImpl;
 
-public class AddRAWFlowTDataMasking extends AbstractJobMigrationTask {
+/**
+ * The class for AddRAWFlowTPatternMasking
+ */
+public class AddRAWFlowMaskingComponents extends AbstractJobMigrationTask {
 
     private ProcessType processType = null;
 
@@ -53,9 +58,11 @@ public class AddRAWFlowTDataMasking extends AbstractJobMigrationTask {
      */
     @Override
     public Date getOrder() {
-        GregorianCalendar gc = new GregorianCalendar(2019, 9, 1, 0, 0, 0);
+        GregorianCalendar gc = new GregorianCalendar(2019, 8, 1, 0, 0, 0);
         return gc.getTime();
     }
+
+
 
     /*
      * (non-Javadoc)
@@ -65,19 +72,35 @@ public class AddRAWFlowTDataMasking extends AbstractJobMigrationTask {
     @Override
     public ExecutionResult execute(Item item) {
         processType = getProcessType(item);
+
         try {
-            IComponentFilter filter = new NameComponentFilter("tDataMasking"); //$NON-NLS-1$
+            IComponentFilter filter = new NameComponentFilter("tPatternMasking"); //$NON-NLS-1$
             IComponentConversion unCheckRaw = new UnCheckRAWFlow();
             IComponentConversion changeMetadataName = new ChangeMetadataName();
-            ModifyComponentsAction.searchAndModify(item, processType, filter,
-                    Arrays.<IComponentConversion> asList(unCheckRaw, changeMetadataName));
+            ModifyComponentsAction
+                    .searchAndModify(item, processType, filter,
+                            Arrays.<IComponentConversion> asList(unCheckRaw, changeMetadataName));
+
+            filter = new NameComponentFilter("tPatternUnmasking"); //$NON-NLS-1$
+            unCheckRaw = new UnCheckRAWFlow();
+            changeMetadataName = new ChangeMetadataName();
+            ModifyComponentsAction
+                    .searchAndModify(item, processType, filter,
+                            Arrays.<IComponentConversion> asList(unCheckRaw, changeMetadataName));
+
+            filter = new NameComponentFilter("tDataMasking"); //$NON-NLS-1$
+            unCheckRaw = new UnCheckRAWFlow();
+            changeMetadataName = new ChangeMetadataName();
+            ModifyComponentsAction
+                    .searchAndModify(item, processType, filter,
+                            Arrays.<IComponentConversion> asList(unCheckRaw, changeMetadataName));
 
             filter = new NameComponentFilter("tDataUnmasking"); //$NON-NLS-1$
             unCheckRaw = new UnCheckRAWFlow();
             changeMetadataName = new ChangeMetadataName();
-            ModifyComponentsAction.searchAndModify(item, processType, filter,
-                    Arrays.<IComponentConversion> asList(unCheckRaw, changeMetadataName));
-
+            ModifyComponentsAction
+                    .searchAndModify(item, processType, filter,
+                            Arrays.<IComponentConversion> asList(unCheckRaw, changeMetadataName));
             return ExecutionResult.SUCCESS_NO_ALERT;
         } catch (Exception e) {
             ExceptionHandler.process(e);
@@ -148,6 +171,9 @@ public class AddRAWFlowTDataMasking extends AbstractJobMigrationTask {
         }
 
         private void removeCustomColumn(MetadataType mt) {
+            if (checkTheColumnExist(mt)) {
+                return;
+            }
             EList<?> columns = mt.getColumn();
             for (Object theColumnObject : columns) {
                 ColumnType theColumn = (ColumnType) theColumnObject;
@@ -165,6 +191,7 @@ public class AddRAWFlowTDataMasking extends AbstractJobMigrationTask {
          * @param metadataName
          */
         private void changeConnection(String metadataName) {
+
             for (Object o : processType.getConnection()) {
                 ConnectionType ct = (ConnectionType) o;
                 if (metadataName.equals(ct.getMetaname()) && OLDCONNECTORNAME.equals(ct.getConnectorName())) {
@@ -180,10 +207,38 @@ public class AddRAWFlowTDataMasking extends AbstractJobMigrationTask {
          * Create new metadata
          */
         private void createNewMetadata(NodeType node, MetadataType mt) {
+
             newMT = EcoreUtil.copy(mt);
             newMT.setConnector(NEWCONNECTORNAME);
             newMT.setName(NEWCONNECTORNAME);
+            node.getMetadata().add(newMT);
+        }
 
+        private boolean checkTheColumnExist(MetadataType mt) {
+            EList<ConnectionTypeImpl> connections = processType.getConnection();
+            EList<NodeTypeImpl> nodes = processType.getNode();
+            String metadataName = mt.getName();
+            for (ConnectionTypeImpl theConnection : connections) {
+                if (metadataName.equals(theConnection.getTarget())) {
+                    for (NodeTypeImpl node : nodes) {
+
+                        EList<MetadataType> previousNodeMetadataTypes = node.getMetadata();
+                        for (MetadataType outputMetadataType : previousNodeMetadataTypes) {
+                            if (outputMetadataType.getName().equals(theConnection.getMetaname())) {
+                                EList<ColumnType> columns = outputMetadataType.getColumn();
+                                for (ColumnType theColumn : columns) {
+                                    if ("ORIGINAL_MARK".equals(theColumn.getName())) { //$NON-NLS-1$
+                                        return true;
+                                    }
+                                }
+                                return false;
+                            }
+                        }
+                    }
+                    return false;
+                }
+            }
+            return false;
         }
     }
 }
